crud_example/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
protobuf==3.19.4
requests==2.27.1
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.3
